requirements.txt
aiogram==3.3.0
python-dotenv
requests
sqlite3
pytz